Create createTablesCode
here is the code for the tables. Now, I gotta get the trigger working for when the admin changes a persons status to infected in the userinfo table, then the locationdata table will change accordingly. Unless, the sql call from the admin does this for you (we will discuss in next scrum).